requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==10.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
